audit

+++ packages/plugins/contracts/test/ConstructorDeploysContract.sol
@@ -2,9 +2,9 @@
 pragma solidity ^0.8.0;
 
 contract DeployedInConstructor {
-    uint public x;
+    uint256 public x;
 
-    constructor(uint _x) {
+    constructor(uint256 _x) {
         x = _x;
     }
 }
@@ -12,11 +12,11 @@ contract DeployedInConstructor {
 contract ConstructorDeploysContract {
     DeployedInConstructor public myContract;
 
-    constructor(uint _x) {
+    constructor(uint256 _x) {
         myContract = new DeployedInConstructor(_x);
     }
 
-    function get() public view returns (uint) {
+    function get() public view returns (uint256) {
         return myContract.x();
     }
 }
