+++ packages/plugins/contracts/test/ContainsStorage.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.0;
 
-import { SimpleStorage } from "./SimpleStorage.sol";
+import {SimpleStorage} from "./SimpleStorage.sol";
 
 library Types {
     enum TestEnum {
@@ -12,7 +12,7 @@ library Types {
 
     type UserDefinedType is uint256;
     type UserDefinedBytes32 is bytes32;
-    type UserDefinedInt is int;
+    type UserDefinedInt is int256;
     type UserDefinedInt8 is int8;
     type UserDefinedUint8 is uint8;
     type UserDefinedBool is bool;
@@ -24,23 +24,24 @@ contract Storage {
         uint128 b;
         uint128 c;
     }
+
     struct ComplexStruct {
         int32 a;
         mapping(uint32 => string) b;
         Types.UserDefinedType c;
     }
 
-    int public immutable immutableInt;
+    int256 public immutable immutableInt;
     int8 public immutable immutableInt8;
-    uint public immutable immutableUint;
+    uint256 public immutable immutableUint;
     uint8 public immutable immutableUint8;
     bool public immutable immutableBool;
     bytes32 public immutable immutableBytes32;
 
     constructor(
-        int _immutableInt,
+        int256 _immutableInt,
         int8 _immutableInt8,
-        uint _immutableUint,
+        uint256 _immutableUint,
         uint8 _immutableUint8,
         bool _immutableBool,
         bytes32 _immutableBytes32
@@ -56,10 +57,10 @@ contract Storage {
     function(uint256) internal pure returns (uint256) internalFunc;
     int8 public minInt8;
     function(uint256) external pure returns (uint256) externalFunc;
-    int public minInt256;
-    int public bigNumberInt256;
+    int256 public minInt256;
+    int256 public bigNumberInt256;
     int8 public bigNumberInt8;
-    uint public bigNumberUint256;
+    uint256 public bigNumberUint256;
     uint8 public bigNumberUint8;
     uint8 public uint8Test;
     bool public boolTest;
@@ -96,15 +97,15 @@ contract Storage {
     SimpleStruct[] public simpleStructDynamicArray;
     mapping(string => string) public stringToStringMapping;
     mapping(string => string) public longStringToLongStringMapping;
-    mapping(string => uint) public stringToUint256Mapping;
+    mapping(string => uint256) public stringToUint256Mapping;
     mapping(string => bool) public stringToBoolMapping;
     mapping(string => address) public stringToAddressMapping;
     mapping(string => SimpleStruct) public stringToStructMapping;
-    mapping(string => uint) public stringToBigNumberUintMapping;
-    mapping(uint => string) public uint256ToStringMapping;
+    mapping(string => uint256) public stringToBigNumberUintMapping;
+    mapping(uint256 => string) public uint256ToStringMapping;
     mapping(uint8 => string) public uint8ToStringMapping;
     mapping(uint128 => string) public uint128ToStringMapping;
-    mapping(int => string) public int256ToStringMapping;
+    mapping(int256 => string) public int256ToStringMapping;
     mapping(int8 => string) public int8ToStringMapping;
     mapping(int128 => string) public int128ToStringMapping;
     mapping(address => string) public addressToStringMapping;
@@ -112,7 +113,7 @@ contract Storage {
     mapping(Types.TestEnum => string) public enumToStringMapping;
     mapping(bytes => string) public bytesToStringMapping;
     mapping(string => mapping(string => string)) public nestedMapping;
-    mapping(uint8 => mapping(string => mapping(address => uint))) public multiNestedMapping;
+    mapping(uint8 => mapping(string => mapping(address => uint256))) public multiNestedMapping;
 
     function getComplexStructMappingVal(uint32 _mappingKey) external view returns (string memory) {
         return complexStruct.b[_mappingKey];
@@ -121,16 +122,16 @@ contract Storage {
 
 contract OtherImmutables {
     Types.UserDefinedType public immutable immutableUserDefinedType;
-    uint public immutable immutableBigNumberUint;
-    int public immutable immutableBigNumberInt;
+    uint256 public immutable immutableBigNumberUint;
+    int256 public immutable immutableBigNumberInt;
     address public immutable immutableAddress;
     Storage public immutable immutableContract;
     Types.TestEnum public immutable immutableEnum;
 
     constructor(
         Types.UserDefinedType _immutableUserDefinedType,
-        uint _immutableBigNumberUint,
-        int _immutableBigNumberInt,
+        uint256 _immutableBigNumberUint,
+        int256 _immutableBigNumberInt,
         address _immutableAddress,
         Storage _immutableContract,
         Types.TestEnum _immutableEnum
